Grammatical tweaks
Minor tweaks to some grammar & punctuation for readability. I've done my best to retain the original wording and meaning as I can, but please feel free to cherry-pick these edits if they're not quite right! :)

Note: The word "Then," (including the comma) at line 108 could be replaced with "That's when" for readability, if my understanding of the intended meaning is correct.